refactor: make the app use FFs on DDB
Ref. metriport/metriport-internal#2840
