+++ packages/api/src/command/internal/get-hie-enabled-feature-flags-status.ts
@@ -1,13 +1,10 @@
-import {
-  getFeatureFlagValueCxValues,
-  CxFeatureFlagStatus,
-} from "@metriport/core/external/aws/app-config";
+import { getFeatureFlagValueCxValues } from "@metriport/core/command/feature-flags";
+import { CxFeatureFlagStatus } from "@metriport/core/command/feature-flags/types";
+import { Config as ConfigCore } from "@metriport/core/util/config";
 import { Config } from "../../shared/config";
 
 export async function getCxFFStatus(cxId: string): Promise<CxFeatureFlagStatus> {
   const region = Config.getAWSRegion();
-  const appId = Config.getAppConfigAppId();
-  const configId = Config.getAppConfigConfigId();
-  const envName = Config.getEnvType();
-  return await getFeatureFlagValueCxValues(region, appId, configId, envName, cxId);
+  const featureFlagsTableName = ConfigCore.getFeatureFlagsTableName();
+  return await getFeatureFlagValueCxValues(region, featureFlagsTableName, cxId);
 }
