+++ packages/infra/lib/api-stack.ts
@@ -142,11 +142,11 @@ export class APIStack extends Stack {
     //-------------------------------------------
     // Application-wide feature flags
     //-------------------------------------------
-    const { appConfigAppId, appConfigConfigId, appConfigEnvId, deploymentStrategyId } =
-      createAppConfigStack({
-        stack: this,
-        props: { config: props.config },
-      });
+    // TODO 2840 remove this
+    createAppConfigStack({
+      stack: this,
+      props: { config: props.config },
+    });
     const { featureFlagsTable } = new FeatureFlagsNestedStack(this, "FeatureFlags", {
       config: props.config,
       alarmAction: slackNotification?.alarmAction,
@@ -348,10 +348,7 @@ export class APIStack extends Stack {
       sandboxSeedDataBucket,
       alarmAction: slackNotification?.alarmAction,
       bedrock: props.config.bedrock,
-      appConfigEnvVars: {
-        appId: appConfigAppId,
-        configId: appConfigConfigId,
-      },
+      featureFlagsTable,
     });
 
     //-------------------------------------------
@@ -447,10 +444,7 @@ export class APIStack extends Stack {
         dashUrl: props.config.dashUrl,
         sentryDsn: props.config.lambdasSentryDSN,
         alarmAction: slackNotification?.alarmAction,
-        appConfigEnvVars: {
-          appId: appConfigAppId,
-          configId: appConfigConfigId,
-        },
+        featureFlagsTable,
         ...props.config.fhirToMedicalLambda,
       });
       fhirToMedicalRecordLambda2 = lambdas.fhirToMedicalRecordLambda2;
@@ -511,12 +505,6 @@ export class APIStack extends Stack {
       searchEndpoint: ccdaSearchDomain.domainEndpoint,
       searchAuth: { userName: ccdaSearchUserName, secret: ccdaSearchSecret },
       searchIndexName: ccdaSearchIndexName,
-      appConfigEnvVars: {
-        appId: appConfigAppId,
-        configId: appConfigConfigId,
-        envId: appConfigEnvId,
-        deploymentStrategyId: deploymentStrategyId,
-      },
       featureFlagsTable,
       cookieStore,
     });
@@ -591,10 +579,7 @@ export class APIStack extends Stack {
       termServerUrl: props.config.termServerUrl,
       apiServiceDnsAddress: apiDirectUrl,
       alarmSnsAction: slackNotification?.alarmAction,
-      appConfigEnvVars: {
-        appId: appConfigAppId,
-        configId: appConfigConfigId,
-      },
+      featureFlagsTable,
     });
 
     // Add ENV after the API service is created
@@ -615,6 +600,7 @@ export class APIStack extends Stack {
     medicalDocumentsBucket.grantReadWrite(documentDownloaderLambda);
     fhirConverterLambda && medicalDocumentsBucket.grantRead(fhirConverterLambda);
 
+    // TODO 2840 remove this
     AppConfigUtils.allowReadConfig({
       scope: this,
       resourceName: "FhirConverterLambda",
@@ -1356,10 +1342,7 @@ export class APIStack extends Stack {
     dashUrl: string;
     sentryDsn: string | undefined;
     alarmAction: SnsAction | undefined;
-    appConfigEnvVars: {
-      appId: string;
-      configId: string;
-    };
+    featureFlagsTable: dynamodb.Table;
   }): { fhirToMedicalRecordLambda2: Lambda } {
     const {
       lambdaLayers,
@@ -1369,7 +1352,7 @@ export class APIStack extends Stack {
       dashUrl,
       alarmAction,
       medicalDocumentsBucket,
-      appConfigEnvVars,
+      featureFlagsTable,
     } = ownProps;
 
     const lambdaTimeout = MAXIMUM_LAMBDA_TIMEOUT.minus(Duration.seconds(5));
@@ -1385,8 +1368,7 @@ export class APIStack extends Stack {
         AXIOS_TIMEOUT_SECONDS: axiosTimeout.toSeconds().toString(),
         MEDICAL_DOCUMENTS_BUCKET_NAME: medicalDocumentsBucket.bucketName,
         PDF_CONVERT_TIMEOUT_MS: CDA_TO_VIS_TIMEOUT.toMilliseconds().toString(),
-        APPCONFIG_APPLICATION_ID: appConfigEnvVars.appId,
-        APPCONFIG_CONFIGURATION_ID: appConfigEnvVars.configId,
+        FEATURE_FLAGS_TABLE_NAME: featureFlagsTable.tableName,
         DASH_URL: dashUrl,
         ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
       },
@@ -1398,12 +1380,7 @@ export class APIStack extends Stack {
       alarmSnsAction: alarmAction,
     });
 
-    AppConfigUtils.allowReadConfig({
-      scope: this,
-      resourceName: "FhirToMrLambda2",
-      resourceRole: fhirToMedicalRecordLambda2.role,
-      appConfigResources: ["*"],
-    });
+    featureFlagsTable.grantReadData(fhirToMedicalRecordLambda2);
 
     medicalDocumentsBucket.grantReadWrite(fhirToMedicalRecordLambda2);
